Split array packer integer handler into header and source files
